Added ammo functionality and pickups
I set it to start the player with 0 ammo since there's no requirement.  I figure we can have them pick up some ammo at the beginning of the level.
Made it so the gun will not fire when ammo is 0 or less
Amount of ammo pickups give can be changed, should probably correlate with size